Support executable names instead of package names
If a name is not the name of an existing package, attempt to locate an
locally installed executable of that name, and use the package it
belongs to instead.

Executables can be specified without a path (this will trigger a $PATH
search), or they may be absolute or relative (will be converted into
absolute ones) path names.